fix: drop gratuitous ull in posit index arithmetic; lns scale() casts to int (#1264) (#1272)
Two independent one-line narrowings (warning-clean Epic #1265):

1. posit_impl.hpp convert_(): `len - nbits - 1ull` in the rounding bit-tests
   (test/anyAfter, lines 385/386) promoted an unsigned index expression to 64 bits
   only to narrow it straight back to unsigned -> -Wconversion. `len`, `nbits` are
   unsigned and test/anyAfter take unsigned, so use 1u. Also cleaned the gratuitous
   `nbits + 1ull` (line 383; it did not warn -- constant that fits -- but is the
   same idiom). len >= nbits + 2 by construction so `len - nbits - 1u` is >= 1 (no
   unsigned wrap).

2. lns_impl.hpp scale(): returned `long(exp)` from an int-returning function ->
   -Wconversion (long is 64-bit on LP64, narrowed to int; no-op on MSVC, hence the
   asymmetric CI reports). int is the scale() convention across the number systems
   (posit/cfloat return int), so the cast was the mistake: use static_cast<int>.

Value-preserving. gcc + clang: the flagged lines are clean under -Wconversion;
posit_addition/multiplication and lns_api/conversion PASS on both.

Resolves #1264

+++ include/sw/universal/number/lns/lns_impl.hpp
@@ -444,7 +444,7 @@ public:
 	constexpr int  scale()  const noexcept {
 		ExponentBlockBinary exp(_block);
 		exp >>= rbits;
-		return long(exp);
+		return static_cast<int>(exp);   // scale() returns int (matches posit/cfloat convention); long(exp) narrowed on LP64
 	}
 	constexpr blockbinary<nbits+2, std::uint32_t, BinaryNumberType::Unsigned> fraction() const noexcept {
 		blockbinary<nbits + 2, std::uint32_t, BinaryNumberType::Unsigned> bb{ 0 };
